# Speed up Jest by using swc instead of tsc