--- a/test/package.json
+++ b/test/package.json
@@ -5,10 +5,12 @@
 	"license": "MIT",
 	"private": true,
 	"scripts": {
-		"_test": "tsc && jest"
+		"_test": "jest"
 	},
 	"devDependencies": {
 		"@moai/core": "*",
+		"@swc/core": "^1.2.174",
+		"@swc/jest": "^0.2.20",
 		"@testing-library/dom": "^8.1.0",
 		"@testing-library/jest-dom": "^5.14.1",
 		"@testing-library/react": "^12.0.0",