# Features
- [x] Speed up Jest tests  by using `@swc/jest` instead of `tsc`.

Results measured on Macbook Pro 2019, running all tests:

|           | Time to finish all tests |
|-----------|--------------------------|
| tsc       | 11.5s                    |
| @swc/jest | 6.0s                     |

Results measured on Macbook Air M1 2020, running all tests:

|           | Time to finish all tests |
|-----------|--------------------------|
| tsc       | 3.68s                    |
| @swc/jest | 2.14s                     |


Conclusion: `@jest/swc` is **91.6%** faster than `tsc` on Macbook with Intel processor, **71.97%** with Macbook with M1 processor.
We can have a more accurate number if the number of test suites and test cases is larger.