@nvh95 2022-05-01
Results measured on Macbook Air M1 2020, `@swc/jest` is **71.97%** faster than `tsc`:

|           | Time to finish all tests |
|-----------|--------------------------|
| tsc       | 3.68s                    |
| @swc/jest | 2.14s                     |